requirements.txt
langgraph>=0.0.19
langchain>=0.0.312
langchain-openai>=0.0.2
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
starlette>=0.27.0
python-dotenv>=1.0.0
pytest>=7.4.3 